Create Lab2b-1
The shaded area within the trapezoid is the area to be painted. The figure inside the trapezoid is a rectangle. Write a program that reads in the length of the two bases and the height of the trapezoid, the length and width of the rectangle ( units are measure in feet), and the cost  per square feet. If the area of the trapezoid is larger, compute the area of the figure that is to be painted.and the cost to paint the shaded area. Display the area to be painted and the cost to paint the shaded area with appropriate labels and accurate to 2 decimal places. If the rectangle is too large, display a message “Cannot compute!”.  Formula to compute area of a trapezoid (a and b are the two bases, h is the height).
// This program will calculate the shaded area to be painted and the cost to paint the shaded area. If